feat: define network selector as action
