+++ src/components/Header/HeaderActions.module.scss
@@ -14,21 +14,16 @@
     width: unset;
   }
 }
-.settings {
-  margin-left: auto;
-}
-.modal {
+.backdrop {
   z-index: 1;
 }
-.widget {
+.dialog {
+  border-radius: 16px;
   background-color: var(--color-surface);
-  height: var(--height);
-  border-radius: 20px;
-  padding-bottom: var(--header-y);
+  position: absolute;
+  width: 100%;
   @media (min-width: 1024px) {
-    padding-bottom: unset;
-    position: absolute;
-    min-width: 240px;
+    max-width: 240px;
     right: 24px;
     top: 68px;
   }
@@ -45,7 +40,8 @@
   margin-right: auto;
 }
 .list {
-  background-color: transparent;
-  padding-top: 16px;
-  padding-bottom: 16px;
+  padding-bottom: var(--header-y);
+  @media (min-width: 1024px) {
+    padding-bottom: unset;
+  }
 }
