+++ src/components/Header/HeaderActions.tsx
@@ -7,18 +7,17 @@ import {
   List,
   ListItem,
   ListItemText,
+  Toggle,
   useMedia
 } from 'ui';
 
 import { Button } from 'components/Button';
 import ConnectMetamask from 'components/ConnectMetamask';
-import CreateMarket from 'components/CreateMarket';
-import Feature from 'components/Feature';
 import Icon from 'components/Icon';
 import Modal from 'components/Modal';
-import Text from 'components/Text';
-import ToggleSwitch from 'components/ToggleSwitch';
 import WalletInfo from 'components/WalletInfo';
+
+import { useNetworks } from 'contexts/networks';
 
 import { useTheme, useAppSelector, usePortal } from 'hooks';
 
@@ -37,14 +36,23 @@ function HeaderActionsWrapper(props: React.PropsWithChildren<{}>) {
   return <Portal {...props} />;
 }
 export default function HeaderActions() {
-  const isDesktop = useMedia('(min-width: 1024px)');
   const isLoggedIn = useAppSelector(state => state.polkamarkets.isLoggedIn);
+  const networks = useNetworks();
   const theme = useTheme();
+  const isDesktop = useMedia('(min-width: 1024px)');
   const [show, setShow] = useState(false);
-  const [height, setHeight] = useState(0);
-  const handleRef = useCallback((ref: HTMLDivElement) => {
-    if (ref !== null) setHeight(ref.getBoundingClientRect().height);
-  }, []);
+  const handleHide = useCallback(() => setShow(false), []);
+  const handleChangeNetwork = useCallback(
+    (event: React.ChangeEvent<HTMLInputElement>) => {
+      const [network] = networks.networks.filter(
+        ({ name }) => name === event.target.value
+      );
+
+      handleHide();
+      networks.changeToNetwork(network);
+    },
+    [handleHide, networks]
+  );
   const isThemeDark = theme.theme === 'dark';
   const HeaderActionsComponent = isDesktop ? 'div' : Container;
   const HeaderActionsRootComponent = isDesktop
@@ -54,11 +62,8 @@ export default function HeaderActions() {
   function handleTheme() {
     theme.setTheme(isThemeDark ? 'light' : 'dark');
   }
-  function handleSettings() {
+  function handleNetworks() {
     setShow(prevShow => !prevShow);
-  }
-  function handleHide() {
-    setShow(false);
   }
 
   return (
@@ -69,71 +74,59 @@ export default function HeaderActions() {
         })}
       >
         {isLoggedIn ? <WalletInfo /> : <ConnectMetamask />}
-        <Button
-          variant="outline"
-          color="default"
-          aria-label="Settings"
-          onClick={handleSettings}
-          className={HeaderActionsClasses.settings}
-        >
-          <Icon name="Gear" />
-        </Button>
+        <div className="pm-c-wallet-info">
+          <Button
+            variant="outline"
+            color="default"
+            aria-label="Switch theme"
+            onClick={handleTheme}
+          >
+            <Icon name={isThemeDark ? 'Sun' : 'Moon'} />
+          </Button>
+          <Button
+            variant="outline"
+            color="default"
+            aria-label="Switch network"
+            onClick={handleNetworks}
+          >
+            <Icon name={networks.network.currency.iconName} />
+          </Button>
+        </div>
       </HeaderActionsComponent>
       <Modal
         disableGutters
+        onHide={handleHide}
         disablePortal={!isDesktop}
         backdrop={!isDesktop}
         show={show}
-        onHide={handleHide}
         className={{
-          root: HeaderActionsClasses.modal
+          backdrop: HeaderActionsClasses.backdrop,
+          dialog: HeaderActionsClasses.dialog
         }}
         {...(!isDesktop && {
-          initial: { y: window.innerHeight },
-          animate: { y: `calc(${window.innerHeight}px - ${height}px)` },
-          exit: { y: window.innerHeight }
+          initial: { bottom: '-100%' },
+          animate: { bottom: 0 },
+          exit: { bottom: '-100%' }
         })}
       >
-        <div ref={handleRef} className={HeaderActionsClasses.widget}>
-          {!isDesktop && (
-            <header className={HeaderActionsClasses.header}>
-              <Text
-                color="light"
-                scale="heading"
-                fontWeight="bold"
-                className={HeaderActionsClasses.title}
-              >
-                Settings
-              </Text>
-              <Adornment edge="end">
-                <Button
-                  size="xs"
-                  variant="ghost"
-                  color="default"
-                  aria-label="Settings"
-                  onClick={handleHide}
-                >
-                  <Icon name="Cross" size="lg" />
-                </Button>
+        <List className={HeaderActionsClasses.list}>
+          {networks.networks.map(network => (
+            <ListItem key={network.id}>
+              <Adornment edge="start">
+                <Icon name={network.currency.iconName} />
               </Adornment>
-            </header>
-          )}
-          <List className={HeaderActionsClasses.list}>
-            <ListItem>
-              <ListItemText>Dark Theme</ListItemText>
-              <Adornment>
-                <ToggleSwitch
-                  name="Switch theme"
-                  checked={isThemeDark}
-                  onChange={handleTheme}
+              <ListItemText>{network.name}</ListItemText>
+              <Adornment edge="end">
+                <Toggle
+                  type="radio"
+                  value={network.name}
+                  checked={network.id === networks.network.id}
+                  onChange={handleChangeNetwork}
                 />
               </Adornment>
             </ListItem>
-            <ListItem>
-              <CreateMarket fullwidth />
-            </ListItem>
-          </List>
-        </div>
+          ))}
+        </List>
       </Modal>
     </HeaderActionsRootComponent>
   );
